fixed windows build, expose include directories, add FILE pointer based model load

+++ test/test_layers.cpp
@@ -495,7 +495,7 @@ inline void debug_tensor_4dxh(const Eigen::Tensor4dXh &x,
                 for (int l = 0; l < x.dimension(3); ++l)
                 {
                     Eigen::half val = x(i, j, k, l);
-                    x_stddev += (val - x_mean) * (val - x_mean);
+                    x_stddev += (static_cast<float>(val) - x_mean) * (static_cast<float>(val) - x_mean);
                 }
             }
         }
